Begin pivot to browser JS
Since 2020, I have come to understand that compared to Javascript in a browser on the web, using the GNU toolkit means more overhead, less compatibility (both cross-platform and future), and less user convenience.

Originally, I had some idea that a user might want to run the configurator on their own system to avoid disclosing personal data to an unknown webpage. I have also learned that not only is this an obscure mentality foreign to most users, but it's also a bit privacy-obsessive.

Hence, we are pivoting to a complete rebuild of the project in browser JavaScript. Honestly, it shouldn't even be that much work ... (fingers crossed)